## Summary

This branch delivers the full Brapi Pro integration as the market data backbone,
a redesigned Market section with fundamentals/FII/macro data, account-to-account
transfers, hardened auth token storage (HttpOnly cookies), routed sub-pages for
Analytics and Simulations, and a portfolio simulator with backtesting and
historical rate suggestion.

## Changes

### Brapi Integration & Market Data Layer
- Add `BrapiPriceUpdateJobService` batch processing for stocks, FIIs, crypto, ETFs, forex pairs, and Tesouro Direto
- Add intraday price tracking (`MarketPriceIntraday` entity + `BrapiIntradayHostedService`) with custom chart period selector
- Add `ProcessFundamentalsBatchAsync`: daily upsert of DY, P/L, P/VP, ROE, marketCap, revenue, net income per asset
- Add `GET /api/market/{ticker}/fii` and `GET /api/market/macro` endpoints with 6h cache
- Add fundamentals-based sort params to `GET /api/market`
- Add `BrapiCleanupJobService` + `RefreshTokenCleanupJobService` for housekeeping

### Market Page Redesign
- Replace standalone search with a dashboard: `MarketIndicatorsStrip` (IBOV/IFIX + forex + macro), `RankingCard`, `MarketAssetRow`
- Add `/market/[ticker]` (asset detail with `FundamentalsPanel` / `FiiPanel`), `/market/ranking`, and `/market/type/[type]` routes
- Remove dead `MarketSearchBar` / `MarketSearchResults`; integrate search into global `GlobalSearch`

### Auth & Security
- Store refresh token as HttpOnly cookie (server-side `Set-Cookie`) — JS can no longer read it
- Access token lives in Zustand memory only; `authStore` no longer persists PII to localStorage
- Add `SetRefreshTokenCookie`/`ClearRefreshTokenCookie` helpers with `HttpOnly`, `Secure`, `SameSite=Strict`
- Move React Query cache persister from localStorage to sessionStorage
- Disable production source maps in `next.config.ts`

### Transactions — Account-to-Account Transfers
- Add `Transfer` transaction type: single DB row with `AccountId` (source) + `DestinationAccountId`
- API: `CreateTransferAsync`/`UpdateTransferAsync`; both accounts must be real and distinct
- Validators extended with `When(Type==Transfer)` rules: destination required, no subcategory/installment/recurrence
- `TransactionDrawer` gains 3-way type toggle and transfer create/edit/detail flow
- `TransactionsList` renders transfers with `ArrowLeftRight` icon, neutral value, and Origem → Destino
- Remove legacy `CreateTransactionModal` and `EditTransactionModal`
- Add `CreateTransactionValidatorTests` transfer suite (4 cases)

### Simulations
- Add `PortfolioSimulator` with backtesting and projection (compound with custom rates)
- Add "Sugerir taxas pelo histórico": period pills → CAGR from local price history → auto-fill rate inputs
- Add `GET /simulation/asset-rate` — annualised CAGR for any local ticker (7D–15A)
- Add `TickerReferenceSelect` in `CompoundInterestSimulator` backed by `useAvailableBenchmarks`
- Convert `/simulations` to routed sub-pages (`/simulations/[slug]`); dynamic DB-backed benchmarks

### Analytics
- Convert `/analytics` to routed sub-pages (`/analytics/[tab]`): gastos, patrimônio, investimentos, projeções
- Add `AnalyticsLayout` with shared `AnalyticsFilterContext` (filter state survives sub-page navigation)
- Fix crash in `ProjectedNetWorthChart` when `historical[]` is empty on first mount

### Investments
- Show Brapi-backed day change in `InvestmentsTable`
- Add accent-insensitive asset search on the register modal
- Fix `UpdateInvestmentAsync` to derive `prevClose` from price history before returning DTO

### DB & Infrastructure
- Add DB indexes: `User.Email`, `RefreshToken.Token`, `Tag(UserId,Name)` unique; `Investment` xmin row-version
- Add `AdminController` for internal ops
- Separate `MarketAsset` from `Investment` with dedicated EF mappings and 4 new migrations

### Refactors / DX
- Extract shared `DateRangePicker`, `MonthRangePicker`, `CategorySelectContent` components
- Add `.claude/specs/` documentation for all major domain areas

+++ apps/api/FinanceControl.Shared/Dtos/Response/Analytics/NetWorthEvolutionItemDto.cs
@@ -4,7 +4,7 @@ namespace FinanceControl.Shared.Dtos.Response.Analytics
     {
         public int Month { get; set; }
         public int Year { get; set; }
-        public int NetWorth { get; set; }
+        public long NetWorth { get; set; }
         public List<AccountBalanceItemDto> Breakdown { get; set; } = [];
     }
 
@@ -12,6 +12,6 @@ namespace FinanceControl.Shared.Dtos.Response.Analytics
     {
         public int AccountId { get; set; }
         public string AccountName { get; set; } = string.Empty;
-        public int Balance { get; set; }
+        public long Balance { get; set; }
     }
 }
